requirements.txt
argon2-cffi==20.1.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
crispy-forms-gds==0.2.2
django-bootstrap4==2.3.1
django-braces==1.14.0
django-crispy-forms==1.10.0
gunicorn==20.1.0
joblib==1.0.1
numpy==1.20.3
Pillow==8.1.0
pycparser==2.20
Pygments==2.7.4
scikit-learn==0.24.2
scipy==1.7.0
six==1.15.0
sklearn==0.0
soupsieve==2.1
threadpoolctl==2.1.0
whitenoise==5.2.0
wincertstore==0.2
